changed api to lowercase - removed version and date in 3 methods
