+++ server/src/services/stakeholder.service.ts
@@ -2,6 +2,7 @@ import Stakeholder from '../schemas/stakeholder.schema';
 import { CustomError } from '@utils/customError';
 import { IStakeholder } from '@interfaces/stakeholder.interface';
 import { StakeholderNotFoundError } from '@utils/errors';
+import { ObjectId } from 'mongoose';
 
 
 //Add new stakeholder
@@ -50,4 +51,40 @@ export const getStakeholdersById = async (
     // Handle error here if needed
     throw new CustomError('Internal Server Error', 500);
   }
+};
+
+export const fetchStakeholders = async (
+  stakeholderIds: ObjectId[],
+): Promise<IStakeholder[]> => {
+  if (stakeholderIds.length > 0) {
+    const stakeholdersResults = await Promise.all(
+      stakeholderIds.map((stakeholderId) => getStakeholdersById(stakeholderId.toString())),
+    );
+    return stakeholdersResults.filter(
+      (stakeholder): stakeholder is IStakeholder => stakeholder !== null,
+    );
+  }
+  return [];
+};
+
+
+export const fetchStakeholdersForSearch = async (
+  stakeholderNames: string[],  
+): Promise<ObjectId[]> => {
+  console.log("Searching for stakeholder names: ", stakeholderNames);
+  
+  if (stakeholderNames && stakeholderNames.length > 0) {
+    // Find stakeholders by names
+    const stakeholders = await Stakeholder.find({
+      type: { $in: stakeholderNames.map(type => new RegExp('^' + type + '$', 'i')) } 
+    }).select('-createdAt -updatedAt -__v');
+
+    console.log("Found stakeholders: ", stakeholders);
+    if (stakeholders.length > 0) {
+      
+      return stakeholders.map(stakeholder => (stakeholder._id as unknown) as ObjectId); // Return the stakeholder IDs
+    }
+  }
+  
+  return []; // no stakeholders found
 };
